consolidate some syscall handler implementations
