+++ Cargo.toml
@@ -42,10 +42,6 @@ build-cli = [
   "dep:cairo-lang-runner",
   "dep:colored",
   "dep:cairo-felt",
-  "dep:keccak",
-  "dep:k256",
-  "dep:p256",
-  "dep:sec1",
 ]
 scarb = ["build-cli", "dep:scarb-ui", "dep:scarb-metadata", "dep:serde_json"]
 with-debug-utils = []
@@ -100,12 +96,12 @@ cairo-lang-runner = { version = "2.6.3", optional = true }
 colored = { version = "2.1.0", optional = true }
 # needed to interface with cairo-lang-*
 cairo-felt = { version = "0.9.1", optional = true }
-keccak = { version = "0.1.3", optional = true }
-k256 = { version = "0.13.3", optional = true }
-p256 = { version = "0.13.2", optional = true }
+keccak = "0.1.3"
+k256 = "0.13.3"
+p256 = "0.13.2"
 scarb-metadata = { git = "https://github.com/software-mansion/scarb.git", rev = "v2.6.3", optional = true }
 scarb-ui = { git = "https://github.com/software-mansion/scarb.git", rev = "v2.6.3", optional = true }
-sec1 = { version = "0.7.3", optional = true }
+sec1 = "0.7.3"
 serde_json = { version = "1.0.117", optional = true }
 
 [dev-dependencies]
