+++ src/lib.rs
@@ -111,6 +111,7 @@ pub mod libfuncs;
 pub mod metadata;
 pub mod module;
 pub mod starknet;
+pub mod starknet_stub;
 pub mod types;
 pub mod utils;
 pub mod values;
